extract common code to shared helper
