+++ src/app/common/app.helpers.ts
@@ -111,3 +111,39 @@ export function changeCopyIcon(event: MouseEvent): void {
         htmlButtonElement.classList.add("clipboard-btn--success");
     }
 }
+
+export function addMarkdownRunButton(sqlQueries: RegExpMatchArray | null, path: string): void {
+    const containerRunButtonElement: HTMLCollectionOf<Element> =
+        document.getElementsByClassName("container-run-button");
+
+    if (sqlQueries?.length && !containerRunButtonElement.length) {
+        const preElements: NodeListOf<Element> = document.querySelectorAll("pre.language-sql");
+        preElements.forEach((preElement: Element, index: number) => {
+            const divElement: HTMLDivElement = document.createElement("div");
+            divElement.classList.add("container-run-button");
+            divElement.style.position = "absolute";
+            divElement.style.top = "10px";
+            divElement.style.right = "65px";
+            const linkElement = document.createElement("a");
+            linkElement.classList.add("markdown-run-button");
+            linkElement.style.padding = "3.6px 16px";
+            linkElement.style.color = "#ffff";
+            linkElement.style.fontSize = "13px";
+            linkElement.style.textDecoration = "none";
+            linkElement.style.backgroundColor = "rgba(255,255, 255, 0.07)";
+            linkElement.style.borderRadius = "4px";
+            linkElement.style.transition = "all 250ms ease-out";
+            linkElement.setAttribute("target", "_blank");
+            linkElement.setAttribute("href", `${path}=${encodeURI(sqlQueries[index])}`);
+            linkElement.addEventListener("mouseover", () => {
+                linkElement.style.backgroundColor = "rgba(255,255, 255, 0.14)";
+            });
+            linkElement.addEventListener("mouseleave", () => {
+                linkElement.style.backgroundColor = "rgba(255,255, 255, 0.07)";
+            });
+            linkElement.innerHTML = "Run";
+            divElement.appendChild(linkElement);
+            preElement.after(divElement);
+        });
+    }
+}
