+++ src/app/dataset-view/additional-components/overview-component/components/readme-section/readme-section.component.ts
@@ -18,6 +18,8 @@ import { DatasetCommitService } from "../../services/dataset-commit.service";
 import { LoggedUserService } from "src/app/auth/logged-user.service";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
 import { DatasetViewTypeEnum } from "src/app/dataset-view/dataset-view.interface";
+import ProjectLinks from "src/app/project-links";
+import { addMarkdownRunButton } from "src/app/common/app.helpers";
 
 @Component({
     selector: "app-readme-section",
@@ -101,42 +103,10 @@ export class ReadmeSectionComponent extends BaseComponent implements OnChanges, 
         if (this.readmeState) {
             // Find all sql queries between ```sql and ```
             const sqlQueries = this.readmeState.match(/(?<=```sql\s+).*?(?=\s+```)/gs);
-            const containerRunButtonElement: HTMLCollectionOf<Element> =
-                document.getElementsByClassName("container-run-button");
-
-            if (sqlQueries?.length && !containerRunButtonElement.length) {
-                const preElements: NodeListOf<Element> = document.querySelectorAll("pre.language-sql");
-                preElements.forEach((preElement: Element, index: number) => {
-                    const divElement: HTMLDivElement = document.createElement("div");
-                    divElement.classList.add("container-run-button");
-                    divElement.style.position = "absolute";
-                    divElement.style.top = "10px";
-                    divElement.style.right = "65px";
-                    const linkElement = document.createElement("a");
-                    linkElement.classList.add("markdown-run-button");
-                    linkElement.style.padding = "3.6px 16px";
-                    linkElement.style.color = "#ffff";
-                    linkElement.style.fontSize = "13px";
-                    linkElement.style.textDecoration = "none";
-                    linkElement.style.backgroundColor = "rgba(255,255, 255, 0.07)";
-                    linkElement.style.borderRadius = "4px";
-                    linkElement.style.transition = "all 250ms ease-out";
-                    linkElement.setAttribute("target", "_blank");
-                    linkElement.setAttribute(
-                        "href",
-                        `/${this.datasetBasics.owner.accountName}/${this.datasetBasics.name}?tab=${DatasetViewTypeEnum.Data}&sqlQuery=${encodeURI(sqlQueries[index])}`,
-                    );
-                    linkElement.addEventListener("mouseover", () => {
-                        linkElement.style.backgroundColor = "rgba(255,255, 255, 0.14)";
-                    });
-                    linkElement.addEventListener("mouseleave", () => {
-                        linkElement.style.backgroundColor = "rgba(255,255, 255, 0.07)";
-                    });
-                    linkElement.innerHTML = "Run";
-                    divElement.appendChild(linkElement);
-                    preElement.after(divElement);
-                });
-            }
+            addMarkdownRunButton(
+                sqlQueries,
+                `/${this.datasetBasics.owner.accountName}/${this.datasetBasics.name}?${ProjectLinks.URL_QUERY_PARAM_TAB}=${DatasetViewTypeEnum.Data}&${ProjectLinks.URL_QUERY_PARAM_SQL_QUERY}`,
+            );
         }
     }
 }
